+++ src/HetrickCV.cpp
@@ -15,6 +15,7 @@ void init(rack::Plugin *p) {
 	p->addModel(modelBitshift);
 	p->addModel(modelBlankPanel);
 	p->addModel(modelBoolean3);
+	p->addModel(modelChaos1Op);
 	p->addModel(modelComparator);
 	p->addModel(modelContrast);
 	p->addModel(modelCrackle);