+++ src/HetrickCV.hpp
@@ -36,6 +36,7 @@ extern Model *modelBinaryNoise;
 extern Model *modelBitshift;
 extern Model *modelBlankPanel;
 extern Model *modelBoolean3;
+extern Model *modelChaos1Op;
 extern Model *modelComparator;
 extern Model *modelContrast;
 extern Model *modelCrackle;