+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # HetrickCV Changelog
 
 ## 2.0.0 Dev
+- Add 1-Op Chaos module.
 - Add Binary Gate module.
 - Add Binary Noise module.
 - Add Gingerbread Chaos module.